create sample schema use volume
