+++ docker-compose.yml
@@ -2,7 +2,8 @@ version: '3'
 services:
   app:
     build: .
+    volumes:
+      - .:/app
     ports:
       - "80:80"
     command: /bin/bash -lc "uvicorn app.main:app --host 0.0.0.0 --port 80"
-    restart: always
